Move cpp files and headers to script dir
