+++ script/cpp/includes/tvgw_filewrite.h
