+++ script/cpp/works cited.txt
@@ -1,0 +1,3 @@
+Thanks for the following site !
+
+http://sky.geocities.jp/kmaedam/directx9/waveform.html
